requirements.txt
fastapi
uvicorn
prisma
pydantic
httpx
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
email-validator
